pyproject.toml
[project]
name = "llm-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "duckduckgo-search>=8.1.1",
    "openai>=2.8.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "requests>=2.32.5",
    "tavily-python>=0.7.13",
    "typer>=0.12.5",
]
[project.scripts]
llm-cli = "llm_cli.cli:app"
